TraceFactory tests
update tests for quiver/streamline with TraceFactory
